% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_genes}
\alias{cluster_genes}
\alias{cluster_samples}
\alias{cluster_both}
\title{Cluster a gene count matrix}
\usage{
cluster_genes(counts_m)

cluster_samples(counts_m)

cluster_both(counts_m)
}
\arguments{
\item{counts_m}{matrix of counts}
}
\value{
matrix
}
\description{
\code{cluster_genes} clusters the matrix by rows (genes)
\code{cluster_samples} clusters the matrix by column (samples)
\code{cluster_both} clusters by both row and column
}
\examples{
counts_m <- matrix(rnorm(25), ncol = 5, 
  dimnames = list(paste("Gene", 1:5, sep = "-"), 
                  paste("sample", 1:5, sep = "-")))

 cluster_genes(counts_m)
 
 cluster_samples(counts_m)
 
 cluster_both(counts_m)
}
