% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterApp}
\alias{clusterApp}
\title{A test shiny app for the cluster module}
\usage{
clusterApp()
}
\value{
a \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
\code{clusterApp()} creates a small test app for testing the \code{\link[=clusterInput]{clusterInput()}} and
\code{\link[=clusterServer]{clusterServer()}} functions. It uses a subset of the package dataset \code{counts},
clusters the counts based on the selected check boxes and displays them
as a table.
}
\examples{

clusterApp()

}
