% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterInput}
\alias{clusterInput}
\title{Create UI components to select whether to cluster a matrix by rows and/or columns}
\usage{
clusterInput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=clusterServer]{clusterServer()}} function.}
}
\value{
a \link[shiny:checkboxGroupInput]{shiny::checkboxGroupInput} object with two check boxes
}
\description{
\code{clusterInput()} produces a checkbox group with check boxes for
clustering by genes (rows) and by samples (columns)
}
\examples{

clusterInput("rnaseqData")

}
