% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPlot.R
\name{countPlotApp}
\alias{countPlotApp}
\title{A test shiny app for the countPlot module}
\usage{
countPlotApp(debug = TRUE)
}
\value{
a \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
\code{countPlotApp()} creates a small test app for testing the \code{\link[=countPlotOutput]{countPlotOutput()}} and
\code{\link[=countPlotServer]{countPlotServer()}} functions. It uses a subset of the package datasets \code{counts},
\code{sampleInfo} and \code{gene_metadata} and create a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} countPlot object.
It also has transform radio buttons to test using that information to change
the colour palette.
}
\examples{

countPlotApp()

}
