% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPlot.R
\name{countPlotInput}
\alias{countPlotInput}
\title{Create UI components to display a count plot for a selected gene}
\usage{
countPlotInput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=countPlotInput]{countPlotInput()}} and \code{\link[=countPlotOutput]{countPlotOutput()}} functions.}
}
\value{
a \code{\link[htmltools:tagList]{htmltools::tagList()}} containing a \code{\link[shiny:selectInput]{shiny::selectInput()}} object
}
\description{
\code{countPlotInput()} produces a \code{\link[shiny:selectInput]{shiny::selectInput()}} control for choosing a gene
to display as a count plot.
}
\examples{

countPlotOutput("rnaseqData")

}
