% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPlot.R
\name{countPlotOutput}
\alias{countPlotOutput}
\title{Create UI components to display a count plot for a selected gene}
\usage{
countPlotOutput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=countPlotServer]{countPlotServer()}} function.}
}
\value{
a \code{\link[htmltools:tagList]{htmltools::tagList()}} containing a \link[shiny:plotOutput]{shiny::plotOutput} object
}
\description{
\code{countPlotOutput()} produces a plotOutput space for a plot
}
\examples{

countPlotOutput("rnaseqData")

}
