% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPlot.R
\name{countPlotServer}
\alias{countPlotServer}
\title{Server function to create a count plot for a given gene}
\usage{
countPlotServer(
  id,
  counts = NULL,
  sample_info = NULL,
  gene_metadata = NULL,
  debug = NULL
)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=countPlotInput]{countPlotInput()}} function.}

\item{counts}{a reactive counts object. Should contain only numeric columns}

\item{sample_info}{a reactive object. Represents the samples and associated
metadata}

\item{gene_metadata}{a reactive object. Contains the metadata for the genes
present in the counts object.}

\item{debug}{Turn on debugging message statements}
}
\value{
a \code{\link[shiny:reactive]{shiny::reactive()}} object which is the count plot
}
\description{
\code{countPlotServer()} creates a count plot for a given gene from the counts and
sample information supplied. The plot is a \link{ggplot2} plot object made using
\code{\link[biovisr:scatterplot_with_fill_and_shape]{biovisr::scatterplot_with_fill_and_shape()}}. The plot is rendered to the
count_plot plotOutput from the \code{countPlotOutput()} function.
}
\examples{

countPlotServer("rnaseq", counts = reactive(rnaseqVis::counts[1:10, 1:20]),
sample_info = reactive(rnaseqVis::sampleInfo[1:20,]),
gene_metadata = reactive(rnaseqVis::gene_metadata[1:10,]))

}
