% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapApp}
\alias{heatmapApp}
\title{A test shiny app for the heatmap module}
\usage{
heatmapApp()
}
\value{
a \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
\code{heatmapApp()} creates a small test app for testing the \code{\link[=heatmapOutput]{heatmapOutput()}} and
\code{\link[=heatmapServer]{heatmapServer()}} functions. It uses a subset of the package datasets \code{counts},
\code{sampleInfo} and \code{gene_metadata} and create a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} heatmap object.
It also has transform radio buttons to test using that information to change
the colour palette.
}
\examples{

heatmapApp()

}
