% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapServer}
\alias{heatmapServer}
\title{Server function to create a heatmap of a count matrix}
\usage{
heatmapServer(
  id,
  counts = NULL,
  sample_info = NULL,
  gene_metadata = NULL,
  transform = NULL,
  debug = NULL
)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=heatmapInput]{heatmapInput()}} function.}

\item{counts}{a reactive counts object. Should contain only numeric columns}
}
\value{
a \code{\link[shiny:reactive]{shiny::reactive()}} object which is the heatmaped counts
}
\description{
\code{heatmapServer()} creates a \link{ggplot2} plot object using \link[biovisr:matrix_heatmap]{biovisr::matrix_heatmap}
from the supplied count matrix, sample info and gene metadata
}
\examples{

heatmapServer("rnaseqData", counts = reactive(rnaseqVis::counts[1:10,1:5]))

}
