% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqVis-package.R
\docType{package}
\name{rnaseqVis-package}
\alias{rnaseqVis}
\alias{rnaseqVis-package}
\title{rnaseqVis: Viewer for RNASeq gene expression data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Richard White \email{rich@buschlab.org} (\href{https://orcid.org/0000-0003-1842-412X}{ORCID})

}
\keyword{internal}
