% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetByGeneIDs.R
\name{SubsetByGeneIDsApp}
\alias{SubsetByGeneIDsApp}
\title{A test shiny app for the SubsetByGeneIDs module}
\usage{
SubsetByGeneIDsApp(debug = FALSE)
}
\arguments{
\item{debug}{logical for turning on debugging messages. default: FALSE}
}
\value{
a \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
\code{SubsetByGeneIDsApp()} creates a small test app for testing the \code{\link[=SubsetByGeneIDsInput]{SubsetByGeneIDsInput()}} and
\code{\link[=SubsetByGeneIDsServer]{SubsetByGeneIDsServer()}} functions. A subset of the returned gene IDs and count data.frames
are displayed in two \code{\link[shiny:renderTable]{shiny::tableOutput()}}s
}
\examples{
SubsetByGeneIDsApp()

SubsetByGeneIDsApp(debug = TRUE)

}
