% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetByGeneIDs.R
\name{SubsetByGeneIDsInput}
\alias{SubsetByGeneIDsInput}
\title{Create UI components to upload a file to use to subset the data}
\usage{
SubsetByGeneIDsInput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=SubsetByGeneIDsServer]{SubsetByGeneIDsServer()}} function.}
}
\value{
a \code{\link[htmltools:tagList]{htmltools::tagList()}} containing a \code{\link[shiny:fileInput]{shiny::fileInput()}} control and
a \code{\link[shiny:actionButton]{shiny::actionButton()}}.
}
\description{
\code{SubsetByGeneIDsInput()} produces upload and reset buttons to provide a file of
gene IDs to use to subset the data with.
}
\examples{
SubsetByGeneIDsInput("geneIds")

}
