% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetByGeneIDs.R
\name{SubsetByGeneIDsOutput}
\alias{SubsetByGeneIDsOutput}
\title{Create UI components to upload a file to use to subset the data}
\usage{
SubsetByGeneIDsOutput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=SubsetByGeneIDsServer]{SubsetByGeneIDsServer()}} function.}
}
\value{
a \code{\link[htmltools:tagList]{htmltools::tagList()}} containing a \code{\link[shinyBS:bsAlert]{shinyBS::bsAlert()}} anchor point
}
\description{
\code{SubsetByGeneIDsOutput()} creates a \code{\link[shinyBS:bsAlert]{shinyBS::bsAlert()}} anchor point to
allow alerting the user if any of the supplied gene IDs are not found in
the data
}
\examples{
SubsetByGeneIDsOutput("geneIds")

}
