% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transformApp}
\alias{transformApp}
\title{A test shiny app for the transform module}
\usage{
transformApp()
}
\value{
a \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
\code{transformApp()} creates a small test app for testing the \code{\link[=transformInput]{transformInput()}} and
\code{\link[=transformServer]{transformServer()}} functions. It uses a subset of the package dataset \code{counts},
transforms the counts based on the selected radio buttons and displays them
as a table.
}
\examples{

transformApp()

}
