% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transformServer}
\alias{transformServer}
\title{Server function to transform a count matrix with a selected function}
\usage{
transformServer(id, counts = NULL)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=transformInput]{transformInput()}} function.}

\item{counts}{a reactive counts object. Should contain only numeric columns}
}
\value{
a list of two \code{\link[shiny:reactive]{shiny::reactive()}} objects
\itemize{
\item counts - the transformed counts
\item transform - the name of the selected transform
}
}
\description{
\code{transformServer()} transforms the supplied counts with the appropriate function
based on the radio buttons selection
}
\examples{

transformServer("rnaseqData", counts = reactive(rnaseqVis::counts[1:10,1:5]))

}
