% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadRNASeq.R
\name{uploadRNASeqApp}
\alias{uploadRNASeqApp}
\title{A test shiny app for the uploadRNASeq module}
\usage{
uploadRNASeqApp()
}
\value{
a \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object
}
\description{
\code{uploadRNASeqApp()} creates a small test app for testing the \code{\link[=uploadRNASeqInput]{uploadRNASeqInput()}} and
\code{\link[=uploadRNASeqServer]{uploadRNASeqServer()}} functions. A subset of the returned sample and count data.frames
are displayed in two \code{\link[shiny:renderTable]{shiny::tableOutput()}}s
}
\examples{
uploadRNASeqApp()
}
