% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadRNASeq.R
\name{uploadRNASeqOutput}
\alias{uploadRNASeqOutput}
\title{Create Output to display uploaded sample and count data}
\usage{
uploadRNASeqOutput(id)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=uploadRNASeqServer]{uploadRNASeqServer()}} function.}
}
\value{
a \code{\link[htmltools:tagList]{htmltools::tagList()}} containing three \code{\link[shiny:renderTable]{shiny::tableOutput()}}s
and two \code{\link[shinyBS:bsAlert]{shinyBS::bsAlert()}} anchor points.
}
\description{
\code{uploadRNASeqOutput()} produces three table outputs for displaying
the data returned by the \code{uploadRNASeqServer()} function. The three
outputs are labelled samples, counts and metadata. There are also two
\code{\link[shinyBS:bsAlert]{shinyBS::bsAlert()}} anchor points to alert the user to missing data in
either the sample or count file
}
\examples{

uploadRNASeqOutput("rnaseqData")

}
