% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadRNASeq.R
\name{uploadRNASeqServer}
\alias{uploadRNASeqServer}
\title{Server function to upload sample and count data files}
\usage{
uploadRNASeqServer(id, debug = FALSE)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=uploadRNASeqInput]{uploadRNASeqInput()}} function.}
}
\value{
a list containing two \code{\link[shiny:reactive]{shiny::reactive()}} objects
\itemize{
\item sampleInfo a data.frame of sample metadata
\item counts a data.frame of RNAseq count data
}
}
\description{
\code{uploadRNASeqServer()} implements uploading a sample file and a count data
file. It also handles using the package test data.
}
\examples{

uploadRNASeqServer("rnaseqData")

}
