% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetByGeneIDs.R
\name{SubsetByGeneIDsServer}
\alias{SubsetByGeneIDsServer}
\title{Server function to upload a file to use to subset the data}
\usage{
SubsetByGeneIDsServer(id, counts = NULL, gene_metadata = NULL, debug = FALSE)
}
\arguments{
\item{id}{namespace id for the UI components. Must match the id provided to the
\code{\link[=SubsetByGeneIDsInput]{SubsetByGeneIDsInput()}} function.}

\item{counts}{a reactive counts object. Should contain only numeric columns}

\item{gene_metadata}{a reactive object. Contains the metadata for the genes
present in the counts object.}

\item{debug}{Turn on debugging message statements}
}
\value{
a \code{\link[shiny:reactive]{shiny::reactive()}} object which is a vector of gene IDs if a file has been uploaded.
If a file hasn't been uploaded or the reset button has been clicked it returns NULL
}
\description{
\code{SubsetByGeneIDsServer()} implements uploading a file of gene IDs to use to
subset the data with. It also deals with resetting the gene IDs with the
reset button.
}
\examples{
SubsetByGeneIDsServer("geneIds")

# turn on debugging
SubsetByGeneIDsServer("geneIds", debug = TRUE)

}
